Adding a test superclass for reading the config properties before all tests.
