+++ src/main/resources/config.properties
@@ -1,14 +1,20 @@
+# Config Properties
+## Log-files
+se.kth.iv1350.log.file.location=data/log
+se.kth.iv1350.log.file.revenue_log=revenue-log.txt
+se.kth.iv1350.log.file.error_log=error-log.txt
+## Discount-Strategies
 se.kth.iv1350.discount_strategy_classname=se.kth.iv1350.integration.pricing.MemberDiscount,se.kth.iv1350.integration.pricing.StudentDiscount,se.kth.iv1350.integration.pricing.Promotion
+## Vat-Strategies
 se.kth.iv1350.vat_strategy.classname=se.kth.iv1350.integration.vat.SwedishVAT
+## Integration
 se.kth.iv1350.database.accounting_db.classname=se.kth.iv1350.integration.AccountingSystemFlatFileDB
 se.kth.iv1350.database.customer_db.classname=se.kth.iv1350.integration.CustomerRegistryFlatFileDB
 se.kth.iv1350.database.inventory_db.classname=se.kth.iv1350.integration.ItemRegistryFlatFileDB
+## Flat-file DB
 se.kth.iv1350.database.file.location=data/db
 se.kth.iv1350.database.file.accounting_db=accounting.csv
 se.kth.iv1350.database.file.customer_db=customers.csv
 se.kth.iv1350.database.file.inventory_db=inventory_items.csv
-se.kth.iv1350.log.file.location=data/log
-se.kth.iv1350.log.file.revenue_log=revenue-log.txt
-se.kth.iv1350.log.file.error_log=error-log.txt
 se.kth.iv1350.database.file.vat_db=vat.csv
 se.kth.iv1350.database.file.csv_delimiter=;
