+++ src/test/se/kth/iv1350/controller/ControllerTest.java
@@ -1,21 +1,18 @@
 package se.kth.iv1350.controller;
 
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import se.kth.iv1350.POSTestSuperClass;
 import se.kth.iv1350.model.*;
 import se.kth.iv1350.integration.*;
-import se.kth.iv1350.startup.Main;
 import se.kth.iv1350.view.EndOfSaleView;
 import se.kth.iv1350.view.RunningSaleView;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.PrintStream;
 import java.time.LocalDateTime;
-import java.util.Properties;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -23,7 +20,7 @@ import static org.junit.jupiter.api.Assertions.*;
    TODO: Check how to send filename to singletons.
    Temp. solution is to use ITEM_ID 0
  */
-class ControllerTest {
+class ControllerTest extends POSTestSuperClass {
     private static final int TEST_ITEM_ID = 0;
     private static final int TEST_QUANTITY = 5;
     private static final String TEST_NAME = "test name";
@@ -36,26 +33,6 @@ class ControllerTest {
     private Controller instance;
     private ByteArrayOutputStream outContent;
     private PrintStream originalSysOut;
-
-    /**
-     * Properties set up base on:
-     * <a href=https://docs.oracle.com/javase/tutorial/essential/environment/sysprop.html>The Java™ Tutorials - System Properties</a>.
-     * If you're having trouble loading the resource file <code>config.properties></code>,
-     * first check that <code>src/test/resources</code>
-     * is correctly configured as a resources directory in your IDE.
-     */
-    @BeforeAll
-    static void setup() {
-        Properties properties = new Properties(System.getProperties());
-        try {
-            InputStream inputStream = Main.class.getClassLoader().getResourceAsStream("config.properties");
-            properties.load(inputStream);
-            System.setProperties(properties);
-        } catch (IOException ex) {
-            System.out.println("Unable to set up configuration");
-            ex.printStackTrace();
-        }
-    }
 
     @BeforeEach
     void setUp() {
