+++ src/test/se/kth/iv1350/model/CashPaymentTest.java
@@ -1,20 +1,18 @@
 package se.kth.iv1350.model;
 
 import org.junit.jupiter.api.*;
+import se.kth.iv1350.POSTestSuperClass;
 import se.kth.iv1350.controller.OperationFailedException;
 import se.kth.iv1350.integration.*;
 import se.kth.iv1350.integration.dto.CustomerDTO;
 import se.kth.iv1350.integration.dto.ItemDTO;
 import se.kth.iv1350.integration.pricing.CustomerType;
-import se.kth.iv1350.startup.Main;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.util.Properties;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class CashPaymentTest {
+class CashPaymentTest extends POSTestSuperClass {
 
     private CashPayment instance;
     private final double PAID = 500;
@@ -24,25 +22,6 @@ class CashPaymentTest {
     private Amount itemPrice = new Amount(10);
     private final int ITEM_ID = 0;
 
-    /**
-     * Properties set up base on:
-     * <a href=https://docs.oracle.com/javase/tutorial/essential/environment/sysprop.html>The Java™ Tutorials - System Properties</a>.
-     * If you're having trouble loading the resource file <code>config.properties></code>,
-     * first check that <code>src/test/resources</code>
-     * is correctly configured as a resources directory in your IDE.
-     */
-    @BeforeAll
-    static void setup() {
-        Properties properties = new Properties(System.getProperties());
-        try {
-            InputStream inputStream = Main.class.getClassLoader().getResourceAsStream("config.properties");
-            properties.load(inputStream);
-            System.setProperties(properties);
-        } catch (IOException ex) {
-            System.out.println("Unable to set up configuration");
-            ex.printStackTrace();
-        }
-    }
     @BeforeEach
     void setUp() throws IOException {
         paidAmount = new Amount(PAID);
@@ -55,7 +34,8 @@ class CashPaymentTest {
         }
         try {
             sale.addItem(ITEM_ID, 1);
-        } catch (ItemRegistryException | ItemNotFoundInItemRegistryException | OperationFailedException e) {
+        } catch (ItemRegistryException | ItemNotFoundInItemRegistryException
+                 | OperationFailedException e) {
             fail("Exception should not have been thrown, " +
                     e.getMessage());
         }
