+++ src/test/se/kth/iv1350/model/CashRegisterTest.java
@@ -1,21 +1,17 @@
 package se.kth.iv1350.model;
 
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import se.kth.iv1350.POSTestSuperClass;
 import se.kth.iv1350.controller.OperationFailedException;
 import se.kth.iv1350.integration.ItemNotFoundInItemRegistryException;
-import se.kth.iv1350.integration.ItemRegistryFlatFileDB;
 import se.kth.iv1350.integration.ItemRegistryException;
-import se.kth.iv1350.startup.Main;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.util.Properties;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class CashRegisterTest {
+class CashRegisterTest extends POSTestSuperClass {
 
     private CashRegister instance;
     private final Amount INITIAL_BALANCE = new Amount(1000);
@@ -25,25 +21,6 @@ class CashRegisterTest {
     private Amount itemPrice = new Amount(10);
     private final int ITEM_ID = 0;
 
-    /**
-     * Properties set up base on:
-     * <a href=https://docs.oracle.com/javase/tutorial/essential/environment/sysprop.html>The Java™ Tutorials - System Properties</a>.
-     * If you're having trouble loading the resource file <code>config.properties></code>,
-     * first check that <code>src/test/resources</code>
-     * is correctly configured as a resources directory in your IDE.
-     */
-    @BeforeAll
-    static void setup() {
-        Properties properties = new Properties(System.getProperties());
-        try {
-            InputStream inputStream = Main.class.getClassLoader().getResourceAsStream("config.properties");
-            properties.load(inputStream);
-            System.setProperties(properties);
-        } catch (IOException ex) {
-            System.out.println("Unable to set up configuration");
-            ex.printStackTrace();
-        }
-    }
     @BeforeEach
     void setUp() throws IOException {
         instance = new CashRegister(INITIAL_BALANCE);
