+++ src/test/se/kth/iv1350/model/SaleTest.java
@@ -1,20 +1,17 @@
 package se.kth.iv1350.model;
 
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import se.kth.iv1350.POSTestSuperClass;
 import se.kth.iv1350.controller.OperationFailedException;
 import se.kth.iv1350.integration.*;
 import se.kth.iv1350.integration.dto.CustomerDTO;
 import se.kth.iv1350.integration.pricing.CustomerType;
-import se.kth.iv1350.startup.Main;
 import se.kth.iv1350.view.EndOfSaleView;
 import se.kth.iv1350.view.RunningSaleView;
 
 import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
 import java.io.PrintStream;
 import java.util.*;
 
@@ -26,7 +23,7 @@ import static org.junit.jupiter.api.Assertions.*;
  * Open issue: What to do with the cash register?
  *
  */
-class SaleTest {
+class SaleTest extends POSTestSuperClass {
     private Sale instance;
     private static final int TEST_ITEM_ID = 0;
     private static final int TEST_QUANTITY = 5;
@@ -43,25 +40,6 @@ class SaleTest {
     private static final int TEST_VAT_GROUP_CODE = 1;
     */
 
-    /**
-     * Properties set up base on:
-     * <a href=https://docs.oracle.com/javase/tutorial/essential/environment/sysprop.html>The Java™ Tutorials - System Properties</a>.
-     * If you're having trouble loading the resource file <code>config.properties></code>,
-     * first check that <code>src/test/resources</code>
-     * is correctly configured as a resources directory in your IDE.
-     */
-    @BeforeAll
-    static void setup() {
-        Properties properties = new Properties(System.getProperties());
-        try {
-            InputStream inputStream = Main.class.getClassLoader().getResourceAsStream("config.properties");
-            properties.load(inputStream);
-            System.setProperties(properties);
-        } catch (IOException ex) {
-            System.out.println("Unable to set up configuration");
-            ex.printStackTrace();
-        }
-    }
     @BeforeEach
     void setUp() throws OperationFailedException {
         instance = new Sale();
